+++ addon/locales/en/components/flexberry-table.js
@@ -1,0 +1,3 @@
+export default {
+  'no-data': 'No data'
+};