+++ addon/locales/en/forms/gis-search-form.js
@@ -1,0 +1,31 @@
+export default {
+  'header': 'Search',
+  'subheader': 'through layer metadata and maps',
+  'keywords': 'Key words',
+  'scale': 'Scale',
+  'object': 'Object',
+  'position': 'Position',
+  'min-long': 'Min longitude',
+  'min-lat': 'Min latitude',
+  'max-long': 'Max longitude',
+  'max-lat': 'Max latitude',
+  'search': 'Search',
+  'results': 'Results',
+  'layer-metadata': {
+    'title': 'Layer metadata',
+    'header': {
+      'name': 'Name',
+      'type': 'Type'
+    }
+  },
+  'maps': {
+    'title': 'Maps',
+    'header': {
+      'name': 'Name',
+      'lat': 'Latitude',
+      'lng': 'Longitude',
+      'zoom': 'Scale',
+      'public': 'Public'
+    }
+  }
+};