+++ addon/locales/ru/components/flexberry-table.js
@@ -1,0 +1,3 @@
+export default {
+  'no-data': 'Нет данных'
+};