+++ addon/locales/ru/components/map-dialogs.js
@@ -1,0 +1,8 @@
+import EditDialog from './map-dialogs/edit';
+import AddDialog from './map-dialogs/add';
+
+export default {
+  'edit': EditDialog,
+
+  'add': AddDialog
+};