+++ addon/locales/ru/components/map-dialogs/edit.js
@@ -1,0 +1,20 @@
+export default {
+  'caption': 'Редактирование карты',
+  'approve-button': {
+    'caption': 'Ок'
+  },
+  'deny-button': {
+    'caption': 'Отмена'
+  },
+  'name-textbox': {
+    'caption': 'Имя карты',
+    'lat': 'Широта',
+    'lng': 'Долгота',
+    'zoom': 'Зум',
+    'public': 'Общая',
+    'description': 'Описание',
+    'keyWords': 'Ключевые слова',
+    'scale': 'Масштаб',
+    'crs': 'Система координат',
+  },
+};