+++ blueprints/ember-flexberry-gis/index.js
@@ -6,25 +6,25 @@ module.exports = {
       // Leaflet.
       {
         name: 'leaflet',
-        target: '1.0.1'
+        target: '1.2.0'
       },
 
       // Leaflet.Proj4.
       {
         name: 'proj4leaflet',
-        target: '1.0.1'
+        target: '1.0.2'
       },
 
-      // Leaflet-WFST.
+      // leaflet-wfst.
       {
-        name: 'Leaflet-WFST',
-        target: 'v2.0.0'
+        name: 'leaflet-wfst',
+        target: '2.0.1-beta.1'
       },
 
       // Leaflet-WMS.
       {
         name: 'leaflet-wms',
-        target: 'v1.1.0'
+        target: '1.2.0-beta.1'
       },
 
       // Leaflet.WMS.
@@ -36,20 +36,13 @@ module.exports = {
       // Leaflet.Editable.
       {
         name: 'leaflet.editable',
-        target: 'master'
-      },
-
-      // Leaflet history.
-      {
-        name: 'leaflet-history',
-        source: 'https://github.com/Flexberry/leaflet-history.git',
-        target: 'd67e852ae405a0ca2f6cd069d2f0ff048d7a5ce8'
+        target: '1.1.0'
       },
 
       // Leaflet.Editable.Measures.
       {
         name: 'leaflet-editable-measures',
-        target: '0.1.0'
+        target: '0.2.0-beta.1'
       },
 
       // Leaflet.Export & dependencies.
@@ -59,23 +52,30 @@ module.exports = {
       },
       {
         name: 'leaflet-export',
-        target: '0.2.0'
+        target: '0.2.1-beta.1'
+      },
+
+      // Leaflet history.
+      {
+        name: 'leaflet-history',
+        source: 'https://github.com/Flexberry/leaflet-history.git',
+        target: 'leaflet-1.2.0'
       },
 
       // Leaflet-switch-scale-control.
       {
         name: 'leaflet-switch-scale-control',
-        target: '0.1.0'
+        target: '0.1.1-beta.1'
       },
 
       // Leaflet.zoomslider.
       {
         name: 'Leaflet.zoomslider',
         source: 'https://github.com/Flexberry/Leaflet.zoomslider.git',
-        target: 'leaflet-1.0.1'
+        target: 'leaflet-1.2.0'
       },
 
-      // Leaflet-MiniMap
+      // Leaflet-MiniMap.
       {
         name: 'leaflet-minimap',
         target: '3.4.0'
@@ -85,32 +85,15 @@ module.exports = {
       {
         name: 'osmtogeojson',
         target: '2.2.12'
-      },
-
-      // Bootstrap slider
-      {
-        name: 'seiyria-bootstrap-slider',
-        target: '~6.0.6'
-      },
-      {
-        name: 'js-beautify',
-        target: '1.6.4'
-      },
-
-      // JQuery-plugin implementing color-picker.
-      {
-        name: 'jquery-minicolors',
-        target: '2.2.6'
       }
     ]).then(function () {
       return _this.addAddonsToProject({
         packages: [{
-            name: 'ember-block-slots',
-            target: '1.1.3'
-          },
-          {
-            name: 'ui-ember-slider',
-            target: '0.5.0'
+            name: 'ember-promise-helpers',
+            target: '1.0.2'
+          }, {
+            name: 'ember-prop-types',
+            target: '2.5.6'
           }
         ]
       });