+++ tests/dummy/app/controllers/application.js
@@ -1,3 +1,7 @@
+/**
+  @module ember-flexberry-gis-dummy
+*/
+
 import Ember from 'ember';
 import config from '../config/environment';
 
@@ -10,6 +14,14 @@ const version = config.APP.version;
   @extends <a href="http://emberjs.com/api/classes/Ember.Controller.html">Ember.Controller</a>
 */
 export default Ember.Controller.extend({
+  /**
+    Service that triggers objectlistview events.
+
+    @property objectlistviewEventsService
+    @type Service
+  */
+  objectlistviewEventsService: Ember.inject.service('objectlistview-events'),
+
   actions: {
     /**
       Toggles application sitemap's side bar.
@@ -17,9 +29,82 @@ export default Ember.Controller.extend({
       @method actions.toggleSidebar
     */
     toggleSidebar() {
-      Ember.$('.ui.sidebar.main.menu').sidebar('toggle');
+      let sidebar = Ember.$('.ui.sidebar.main.menu');
+      let objectlistviewEventsService = this.get('objectlistviewEventsService');
+      sidebar.sidebar({
+        closable: false,
+        dimPage: false,
+        onHide: function() {
+          Ember.$('.sidebar.icon.text-menu-show').removeClass('hidden');
+          Ember.$('.sidebar.icon.text-menu-hide').addClass('hidden');
+        },
+        onHidden: function() {
+          objectlistviewEventsService.updateWidthTrigger();
+        },
+        onShow: function() {
+          objectlistviewEventsService.updateWidthTrigger();
+        }
+      }).sidebar('toggle');
+
+      if (Ember.$('.inverted.vertical.main.menu').hasClass('visible')) {
+        Ember.$('.sidebar.icon.text-menu-show').removeClass('hidden');
+        Ember.$('.sidebar.icon.text-menu-hide').addClass('hidden');
+        Ember.$('.bgw-opacity').addClass('hidden');
+      } else {
+        Ember.$('.sidebar.icon.text-menu-show').addClass('hidden');
+        Ember.$('.sidebar.icon.text-menu-hide').removeClass('hidden');
+        Ember.$('.bgw-opacity').removeClass('hidden');
+      }
+
+      if (Ember.$('.inverted.vertical.main.menu').hasClass('visible')) {
+        Ember.$('.full.height').css({ transition: 'width 0.45s ease-in-out 0s', width: '100%' });
+      } else {
+        Ember.$('.full.height').css({ transition: 'width 0.3s ease-in-out 0s', width: 'calc(100% - ' + sidebar.width() + 'px)' });
+      }
+    },
+
+    /**
+      Toggles application sitemap's side bar in mobile view.
+
+      @method actions.toggleSidebarMobile
+    */
+    toggleSidebarMobile() {
+      let sidebar = Ember.$('.ui.sidebar.main.menu');
+      let objectlistviewEventsService = this.get('objectlistviewEventsService');
+      sidebar.sidebar({
+        onHide: function() {
+          Ember.$('.sidebar.icon.text-menu-show').removeClass('hidden');
+          Ember.$('.sidebar.icon.text-menu-hide').addClass('hidden');
+        },
+        onHidden: function() {
+          objectlistviewEventsService.updateWidthTrigger();
+        },
+        onShow: function() {
+          objectlistviewEventsService.updateWidthTrigger();
+        }
+      }).sidebar('toggle');
+
+      if (Ember.$('.inverted.vertical.main.menu').hasClass('visible')) {
+        Ember.$('.sidebar.icon.text-menu-show').removeClass('hidden');
+        Ember.$('.sidebar.icon.text-menu-hide').addClass('hidden');
+        Ember.$('.bgw-opacity').addClass('hidden');
+      } else {
+        Ember.$('.sidebar.icon.text-menu-show').addClass('hidden');
+        Ember.$('.sidebar.icon.text-menu-hide').removeClass('hidden');
+        Ember.$('.bgw-opacity').removeClass('hidden');
+      }
     }
   },
+
+  /**
+    Flag: indicates that form to which controller is related designed for acceptance tests &
+    all additional markup in application.hbs mustn't be rendered.
+
+    @property isInAcceptanceTestMode
+    @type Boolean
+    @default false
+  */
+  isInAcceptanceTestMode: false,
 
   /**
     Currernt addon version.
@@ -35,11 +120,11 @@ export default Ember.Controller.extend({
     @property addonVersionHref
     @type String
   */
-  addonVersionHref: Ember.computed('addonVersion', function () {
+  addonVersionHref: Ember.computed('addonVersion', function() {
     let addonVersion = this.get('addonVersion');
     let commitSha = addonVersion.split('+')[1];
 
-    return 'https://github.com/Flexberry/ember-flexberry-gis/commit/' + commitSha;
+    return 'https://github.com/Flexberry/ember-flexberry/commit/' + commitSha;
   }),
 
   /**
@@ -48,7 +133,7 @@ export default Ember.Controller.extend({
     @property browserIsInternetExplorer
     @type Boolean
   */
-  browserIsInternetExplorer: Ember.computed(function () {
+  browserIsInternetExplorer: Ember.computed(function() {
     let userAgent = window.navigator.userAgent;
 
     return userAgent.indexOf('MSIE ') > 0 || userAgent.indexOf('Trident/') > 0 || userAgent.indexOf('Edge/') > 0;
@@ -62,6 +147,16 @@ export default Ember.Controller.extend({
     @default ['ru', 'en']
   */
   locales: ['ru', 'en'],
+
+  /**
+    Handles changes in userSettingsService.isUserSettingsServiceEnabled.
+
+    @method _userSettingsServiceChanged
+    @private
+  */
+  _userSettingsServiceChanged: Ember.observer('userSettingsService.isUserSettingsServiceEnabled', function() {
+    this.get('target.router').refresh();
+  }),
 
   /**
     Initializes controller.
@@ -112,63 +207,17 @@ export default Ember.Controller.extend({
           caption: i18n.t('forms.application.sitemap.gis.maps.caption'),
           title: i18n.t('forms.application.sitemap.gis.maps.title'),
           children: null
-        },
-        {
+        }, {
           link: 'new-platform-flexberry-g-i-s-layer-metadata-l',
           caption: i18n.t('forms.application.sitemap.gis.map-metadata.caption'),
           title: i18n.t('forms.application.sitemap.gis.map-metadata.title'),
           children: null
+        }, {
+          link: 'gis-search-form',
+          caption: i18n.t('forms.application.sitemap.gis.gis-search-form.caption'),
+          title: i18n.t('forms.application.sitemap.gis.gis-search-form.title'),
+          children: null
         }]
-      }, {
-        link: null,
-        caption: i18n.t('forms.application.sitemap.components-examples.caption'),
-        title: i18n.t('forms.application.sitemap.components-examples.title'),
-        children: [{
-          link: null,
-          caption: i18n.t('forms.application.sitemap.components-examples.flexberry-button.caption'),
-          title: i18n.t('forms.application.sitemap.components-examples.flexberry-button.title'),
-          children: [{
-            link: 'components-examples/flexberry-button/settings-example',
-            caption: i18n.t('forms.application.sitemap.components-examples.flexberry-button.settings-example.caption'),
-            title: i18n.t('forms.application.sitemap.components-examples.flexberry-button.settings-example.title'),
-            children: null
-          }]
-        }, {
-          link: null,
-          caption: i18n.t('forms.application.sitemap.components-examples.flexberry-ddau-checkbox.caption'),
-          title: i18n.t('forms.application.sitemap.components-examples.flexberry-ddau-checkbox.title'),
-          children: [{
-            link: 'components-examples/flexberry-ddau-checkbox/settings-example',
-            caption: i18n.t('forms.application.sitemap.components-examples.flexberry-ddau-checkbox.settings-example.caption'),
-            title: i18n.t('forms.application.sitemap.components-examples.flexberry-ddau-checkbox.settings-example.title'),
-            children: null
-          }]
-        }, {
-          link: null,
-          caption: i18n.t('forms.application.sitemap.components-examples.flexberry-maplayers.caption'),
-          title: i18n.t('forms.application.sitemap.components-examples.flexberry-maplayers.title'),
-          children: [{
-            link: 'components-examples/flexberry-maplayers/settings-example',
-            caption: i18n.t('forms.application.sitemap.components-examples.flexberry-maplayers.settings-example.caption'),
-            title: i18n.t('forms.application.sitemap.components-examples.flexberry-maplayers.settings-example.title'),
-            children: null
-          }]
-        }, {
-          link: null,
-          caption: i18n.t('forms.application.sitemap.components-examples.flexberry-tree.caption'),
-          title: i18n.t('forms.application.sitemap.components-examples.flexberry-tree.title'),
-          children: [{
-            link: 'components-examples/flexberry-tree/settings-example',
-            caption: i18n.t('forms.application.sitemap.components-examples.flexberry-tree.settings-example.caption'),
-            title: i18n.t('forms.application.sitemap.components-examples.flexberry-tree.settings-example.title'),
-            children: null
-          }]
-        }]
-      }, {
-        link: null,
-        caption: i18n.t('forms.application.sitemap.integration-examples.caption'),
-        title: i18n.t('forms.application.sitemap.integration-examples.title'),
-        children: null
       }]
     };
   })