+++ tests/dummy/app/locales/en/forms/new-platform-flexberry-g-i-s-map-e.js
@@ -1,9 +1,0 @@
-export default {
-  caption: 'NewPlatformFlexberryGISMapE',
-  'name-caption': 'name',
-  'lat-caption': 'lat',
-  'lng-caption': 'lng',
-  'zoom-caption': 'zoom',
-  'public-caption': 'public',
-  'coordinateReferenceSystem-caption': 'coordinateReferenceSystem'
-};