+++ tests/dummy/app/locales/en/translations.js
@@ -2,11 +2,6 @@ import Ember from 'ember';
 import EmberFlexberryTranslations from 'ember-flexberry/locales/en/translations';
 import EmberFlexberryGisTranslations from 'ember-flexberry-gis/locales/en/translations';
 
-import NewPlatformFlexberryGISLayerMetadataLForm from './forms/new-platform-flexberry-g-i-s-layer-metadata-l';
-import NewPlatformFlexberryGISMapLForm from './forms/new-platform-flexberry-g-i-s-map-l';
-import NewPlatformFlexberryGISLayerMetadataEForm from './forms/new-platform-flexberry-g-i-s-layer-metadata-e';
-import NewPlatformFlexberryGISMapEForm from './forms/new-platform-flexberry-g-i-s-map-e';
-
 const translations = {};
 Ember.$.extend(true, translations, EmberFlexberryTranslations, EmberFlexberryGisTranslations);
 
@@ -30,6 +25,12 @@ Ember.$.extend(true, translations, {
           },
           'user-settings-service-checkbox': {
             'caption': 'Use user settings service'
+          },
+          'show-menu': {
+            'caption': 'Show menu'
+          },
+          'hide-menu': {
+            'caption': 'Hide menu'
           },
           'language-dropdown': {
             'caption': 'Application language',
@@ -77,6 +78,10 @@ Ember.$.extend(true, translations, {
           'map-metadata': {
             'caption': 'Мetadata',
             'title': ''
+          },
+          'gis-search-form': {
+            'caption': 'Search',
+            'title': 'Search through layer metadata and maps'
           }
         },
         'components-examples': {
@@ -132,16 +137,6 @@ Ember.$.extend(true, translations, {
     },
 
     'components-examples': {
-      'flexberry-button': {
-        'settings-example': {
-          'caption': 'Settings example for flexberry-button'
-        }
-      },
-      'flexberry-ddau-checkbox': {
-        'settings-example': {
-          'caption': 'Settings example for flexberry-ddau-checkbox'
-        }
-      },
       'flexberry-maplayers': {
         'settings-example': {
           'caption': 'Settings example for flexberry-maplayers',
@@ -150,22 +145,11 @@ Ember.$.extend(true, translations, {
           'json-layers-latest-clicked-layer-placeholder': 'Click on any layer to display it\'s settings'
         }
       },
-      'flexberry-tree': {
-        'settings-example': {
-          'caption': 'Settings example for flexberry-tree',
-          'json-tree-tab-caption': 'JSON-object-defined tree',
-          'json-tree-latest-clicked-node-caption': 'Latest clicked tree node settings',
-          'json-tree-latest-clicked-node-placeholder': 'Click on any tree node to display it\'s settings'
-        }
-      }
     },
 
-    'integration-examples': {
+    'maps': {
+      'caption': 'Maps'
     },
-    'new-platform-flexberry-g-i-s-layer-metadata-l': NewPlatformFlexberryGISLayerMetadataLForm,
-    'new-platform-flexberry-g-i-s-map-l': NewPlatformFlexberryGISMapLForm,
-    'new-platform-flexberry-g-i-s-layer-metadata-e': NewPlatformFlexberryGISLayerMetadataEForm,
-    'new-platform-flexberry-g-i-s-map-e': NewPlatformFlexberryGISMapEForm,
     'map': {
       'caption': 'Map',
       'treeviewbuttontooltip': 'Show tree view',