+++ tests/dummy/app/locales/ru/forms/new-platform-flexberry-g-i-s-layer-metadata-e.js
@@ -1,7 +1,0 @@
-export default {
-  caption: '',
-  'name-caption': 'Name',
-  'type-caption': 'Type',
-  'coordinateReferenceSystem-caption': 'Coordinate reference system',
-  'settings-caption': 'Settings'
-};