+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Added
+* Add Semantic UI/Flexberry themes support.
+* Add `flexberry-edit-layermap` component to unify layers and layers-metadata editing.
+* Add `gis-search-form` making available search through maps and layer-metadata.
+* Add `edit` map dialog allowing map properties editing.
+
+## [0.2.1] - 2017-09-04
+### Removed
+* Remove `isDevelopingAddon` property from release.
 
 ## [0.2.0] - 2017-08-30
 ### Added